+++ addOns/authhelper/CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Include step for errors in the authentication diagnostics.
 - Browser based authentication to also support HTTP basic authentication for Firefox.
 - Verification rule to improve detection.
+- Add support for Microsoft login in Browser Based Authentication.
 
 ### Fixed
 - Do not fail the authentication on diagnostic errors.